split api
