+++ server.js
@@ -8,7 +8,8 @@ var bodyParser = require('body-parser');
 var http = require('http');
 var WebSocketServer = require('ws').Server;
 
-// var api = require('./api');
+var figures = require('./hub/figures');
+var snapshots = require('./hub/snapshots');
 
 var knexConfig = require('./knexfile');
 var port = process.env.PORT || 5000;
@@ -49,19 +50,6 @@ app.use(express.static(path.join(__dirname, 'notepad')));
 app.use('/figures', express.static(path.join(__dirname, 'uploads')));
 app.use('/fonts', express.static(path.join(__dirname, 'node_modules/font-awesome/fonts')));
 
-
-// Error handler
-// app.use(errorHandler);
-
-// Connect http api
-// app.use('/api', api(store));
-
-
-// function errorHandler(err, req, res) {
-//   res.status(500);
-//   res.send(err.message);
-// }
-
 // Connect Substance
 // ----------------
 
@@ -71,6 +59,11 @@ var wss = new WebSocketServer({ server: httpServer });
 var hub = new CollabHub(wss, backend);
 // Adds http routes that CollabHub implements
 hub.addRoutes(app);
+
+snapshots.connect(backend);
+snapshots.addRoutes(app);
+
+figures.addRoutes(app);
 
 // Delegate http requests to express app
 httpServer.on('request', app);
